move value translator to speedy
