+++ sdk/daml-lf/ide-ledger/src/main/scala/com/digitalasset/daml/lf/script/IdeLedgerRunner.scala
@@ -257,7 +257,7 @@ private[lf] object IdeLedgerRunner {
         requireContractIdSuffix = config.requireSuffixedGlobalContractId,
       )
     def translateValue(typ: Ast.Type, value: Value): Result[SValue] =
-      valueTranslator.strictTranslateValue(typ, value) match {
+      valueTranslator.translateValue(typ, value) match {
         case Left(err) => ResultError(err)
         case Right(sv) => ResultDone(sv)
       }
